consulta jpql named-query no orm.xml

+++ src/main/java/com/algaworks/algafood/domain/model/Restaurante.java
@@ -26,7 +26,7 @@ public class Restaurante {
 	private String nome;
 	
 	@Column(name="taxa_frete")
-	private Double taxaFrete;
+	private BigDecimal taxaFrete;
 	
 	@ManyToOne
 	private Cozinha cozinha;
